Prepare kubectl
